+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * oauth: AppleOAuthClient updated and tested
 * oauth: JWT helper class introduced
 * jdbc: introduced db.upsertBatch()
+* jobs: runOnce() introduced for convenience
 
 # 1.6.4
 * core: added common value types for Email, Phone, and Password, also StringValue base class