+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * oauth: JWT helper class introduced
 * jdbc: introduced db.upsertBatch()
 * jobs: runOnce() introduced for convenience
+* json: throw more descriptive errors from JsonNode.get() in case of missing keys
 
 # 1.6.4
 * core: added common value types for Email, Phone, and Password, also StringValue base class